requirements.txt
streamlit
openai
pymongo
faiss-cpu
numpy
Pillow
dlib
face-recognition
